refactor(ios): VeepooSDK.swift is now AsyncFunction wiring only
Every heavy inline AsyncFunction body has moved into a
VeepooSDKModule+<Feature>.swift extension (Session / Scan / Connect /
Battery / PersonalInfo / DeviceInfo / SocialMsgIO / Language), plus
stop-handlers for the six realtime tests now live next to their start
counterparts. The 49-row VeepooEvent enum and the matching SCREAMING
aliases moved out to VeepooEvents.swift; ConnectionState and
PermissionDelegate each got their own file.

VeepooSDK.swift is now 306 lines: imports, Module subclass state +
helpers, the Events(...) registration, and one-line AsyncFunction
wrappers that dispatch into the per-feature handlers. The slight
overage against the issue's ~300-line target is structural — Expo's
ModuleDefinitionBuilder result builder only supports a flat
buildBlock variadic, so each AsyncFunction call must stay in the
definition body — and is documented at the top of the file.

The bridge-contract verifier was updated to read VeepooEvents.swift
instead of slicing the top of VeepooSDK.swift. JS suite still
931 / 931.

Fixes #194
