+++ ios/VeepooSDK/VeepooEvents.swift
@@ -1,0 +1,111 @@
+import Foundation
+
+// MARK: - 事件常量
+enum VeepooEvent {
+  static let deviceFound = "deviceFound"
+  static let deviceConnected = "deviceConnected"
+  static let deviceDisconnected = "deviceDisconnected"
+  static let deviceConnectStatus = "deviceConnectStatus"
+  static let connectionStatusChanged = "connectionStatusChanged"
+  static let deviceReady = "deviceReady"
+  static let bluetoothStateChanged = "bluetoothStateChanged"
+  static let deviceFunction = "deviceFunction"
+  static let deviceVersion = "deviceVersion"
+  static let passwordData = "passwordData"
+  static let batteryData = "batteryData"
+  static let heartRateTestResult = "heartRateTestResult"
+  static let bloodPressureTestResult = "bloodPressureTestResult"
+  static let bloodOxygenTestResult = "bloodOxygenTestResult"
+  static let temperatureTestResult = "temperatureTestResult"
+  static let stressData = "stressData"
+  static let bloodGlucoseData = "bloodGlucoseData"
+  static let sleepData = "sleepData"
+  static let sportStepData = "sportStepData"
+  static let readOriginProgress = "readOriginProgress"
+  static let readOriginComplete = "readOriginComplete"
+  static let originFiveMinuteData = "originFiveMinuteData"
+  static let originHalfHourData = "originHalfHourData"
+  static let originSpo2Data = "originSpo2Data"
+  static let socialMsgData = "socialMsgData"
+  static let findDeviceState = "findDeviceState"
+  static let firmwareDfuProgress = "firmwareDfuProgress"
+  static let cameraShutter = "cameraShutter"
+  static let musicRemoteCommand = "musicRemoteCommand"
+  static let deviceBTStateChanged = "deviceBTStateChanged"
+  static let deviceSosTriggered = "deviceSosTriggered"
+  static let customSettingsData = "customSettingsData"
+  static let healthRemindData = "healthRemindData"
+  static let apneaRemindData = "apneaRemindData"
+  static let sportModeData = "sportModeData"
+  static let bloodAnalysisTestResult = "bloodAnalysisTestResult"
+  static let gsrTestResult = "gsrTestResult"
+  static let exerciseSessionData = "exerciseSessionData"
+  static let accurateSleepData = "accurateSleepData"
+  static let storedTemperatureData = "storedTemperatureData"
+  static let storedBloodGlucoseData = "storedBloodGlucoseData"
+  static let storedHrvData = "storedHrvData"
+  static let storedEcgData = "storedEcgData"
+  static let storedBodyCompositionData = "storedBodyCompositionData"
+  static let pttTestResult = "pttTestResult"
+  static let pttStateChanged = "pttStateChanged"
+  static let error = "error"
+}
+
+// Module-private aliases used at call sites to keep `sendEvent(...)` and the
+// `Events(...)` registration compact. Kept here so adding an event is one row
+// in `VeepooEvent` plus one alias here.
+let DEVICE_FOUND = VeepooEvent.deviceFound
+let DEVICE_CONNECTED = VeepooEvent.deviceConnected
+let DEVICE_DISCONNECTED = VeepooEvent.deviceDisconnected
+let DEVICE_CONNECT_STATUS = VeepooEvent.deviceConnectStatus
+let CONNECTION_STATUS_CHANGED = VeepooEvent.connectionStatusChanged
+let DEVICE_READY = VeepooEvent.deviceReady
+let BLUETOOTH_STATE_CHANGED = VeepooEvent.bluetoothStateChanged
+let DEVICE_FUNCTION = VeepooEvent.deviceFunction
+let DEVICE_VERSION = VeepooEvent.deviceVersion
+let PASSWORD_DATA = VeepooEvent.passwordData
+let BATTERY_DATA = VeepooEvent.batteryData
+let HEART_RATE_TEST_RESULT = VeepooEvent.heartRateTestResult
+let BLOOD_PRESSURE_TEST_RESULT = VeepooEvent.bloodPressureTestResult
+let BLOOD_OXYGEN_TEST_RESULT = VeepooEvent.bloodOxygenTestResult
+let TEMPERATURE_TEST_RESULT = VeepooEvent.temperatureTestResult
+let STRESS_DATA = VeepooEvent.stressData
+let BLOOD_GLUCOSE_DATA = VeepooEvent.bloodGlucoseData
+let SLEEP_DATA = VeepooEvent.sleepData
+let SPORT_STEP_DATA = VeepooEvent.sportStepData
+let READ_ORIGIN_PROGRESS = VeepooEvent.readOriginProgress
+let READ_ORIGIN_COMPLETE = VeepooEvent.readOriginComplete
+let ORIGIN_FIVE_MINUTE_DATA = VeepooEvent.originFiveMinuteData
+let ORIGIN_HALF_HOUR_DATA = VeepooEvent.originHalfHourData
+let ORIGIN_SPO2_DATA = VeepooEvent.originSpo2Data
+let SOCIAL_MSG_DATA = VeepooEvent.socialMsgData
+let FIND_DEVICE_STATE = VeepooEvent.findDeviceState
+let FIRMWARE_DFU_PROGRESS = VeepooEvent.firmwareDfuProgress
+let CAMERA_SHUTTER = VeepooEvent.cameraShutter
+let MUSIC_REMOTE_COMMAND = VeepooEvent.musicRemoteCommand
+let DEVICE_BT_STATE_CHANGED = VeepooEvent.deviceBTStateChanged
+let DEVICE_SOS_TRIGGERED = VeepooEvent.deviceSosTriggered
+let CUSTOM_SETTINGS_DATA = VeepooEvent.customSettingsData
+let HEALTH_REMIND_DATA = VeepooEvent.healthRemindData
+let APNEA_REMIND_DATA = VeepooEvent.apneaRemindData
+let SPORT_MODE_DATA = VeepooEvent.sportModeData
+let BLOOD_ANALYSIS_TEST_RESULT = VeepooEvent.bloodAnalysisTestResult
+let GSR_TEST_RESULT = VeepooEvent.gsrTestResult
+let EXERCISE_SESSION_DATA = VeepooEvent.exerciseSessionData
+let ACCURATE_SLEEP_DATA = VeepooEvent.accurateSleepData
+let STORED_TEMPERATURE_DATA = VeepooEvent.storedTemperatureData
+let STORED_BLOOD_GLUCOSE_DATA = VeepooEvent.storedBloodGlucoseData
+let STORED_HRV_DATA = VeepooEvent.storedHrvData
+let STORED_ECG_DATA = VeepooEvent.storedEcgData
+let STORED_BODY_COMPOSITION_DATA = VeepooEvent.storedBodyCompositionData
+let PTT_TEST_RESULT = VeepooEvent.pttTestResult
+let PTT_STATE_CHANGED = VeepooEvent.pttStateChanged
+let ALARM_DATA = "alarmData"
+let HRV_TEST_RESULT = "hrvTestResult"
+let ECG_TEST_RESULT = "ecgTestResult"
+let FATIGUE_TEST_RESULT = "fatigueTestResult"
+let BREATHING_TEST_RESULT = "breathingTestResult"
+let BODY_COMPOSITION_TEST_RESULT = "bodyCompositionTestResult"
+let CONTACTS_DATA = "contactsData"
+let SOS_CALL_TIMES_DATA = "sosCallTimesData"
+let ERROR = VeepooEvent.error
